MainApp/requirements.txt
base64
json
opencv-python
numpy
kafka-python
paho-mqtt
ultralytics
configparser
time
